orca.el (orca-handler-current-buffer): Add a workaround for capturing
Since I started using `org-roam', `orca-handler-current-buffer' was
working on-and-off.  I had to do `server-force-delete' and
`server-start' to make it work again.  With this workaround (to a
mechanism which was flaky in the first place), it seems to work again.

I wish I had a better solution, especially since with `org-roam' I'd
like to do recursive captures, i.e. `org-capture' a like from Firefox
into a capture buffer of `org-roam'.